refactor: extract github and project registry routes
